+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ts-linq/cli
 
+## 1.3.29
+
+### Patch Changes
+
+- Updated dependencies
+  - @ts-linq/core@3.2.0
+  - @ts-linq/migrations@2.6.23
+
 ## 1.3.28
 
 ### Patch Changes